Track armour values more consistently
Before, I was tracking armour values via a threshold (the top X HP is counted as armoured).
This was weird and unintuitive, and just makes tracking things a bit of a pain.

This commit simplifies armour to be tracked alongside HP as a value (though it's really part of the HP).

Also it splits the armour damage reduction into its own stat (Armour Class)