Allow allocating skill nodes along either direction of a connection
Nodes in a skill tree only have their connections marked in one direction (atm, the later node refs the prereq)
This is done for data consistency reasons because it's easy to accidentally miss a direction.
This commit adds support in the PlayerSerice for tracking the inverse connections,
to allow allocation even if there isn't an explicitly defined connection.